+++ src/dto/site.js
@@ -11,7 +11,6 @@
  */
 
 import { createSite } from '@adobe/spacecat-shared-data-access/src/models/site.js';
-import AuditConfig from '@adobe/spacecat-shared-data-access/src/models/site/audit-config.js';
 import { Config } from '@adobe/spacecat-shared-data-access/src/models/site/config.js';
 import { AuditDto } from './audit.js';
 
@@ -36,7 +35,6 @@ export const SiteDto = {
       isLive: jsonObject.isLive,
       createdAt: jsonObject.createdAt,
       updatedAt: jsonObject.updatedAt,
-      auditConfig: AuditConfig.fromDynamoItem(jsonObject.auditConfig),
       config: jsonObject.config,
       audits: jsonObject.audits ? jsonObject.audits.map((audit) => AuditDto.fromJson(audit)) : [],
     };
@@ -68,7 +66,6 @@ export const SiteDto = {
     isLiveToggledAt: site.getIsLiveToggledAt(),
     createdAt: site.getCreatedAt(),
     updatedAt: site.getUpdatedAt(),
-    auditConfig: AuditConfig.toDynamoItem(site.getAuditConfig()),
     config: Config.toDynamoItem(site.getConfig()),
     ...(site.getAudits().length > 0
       && { audits: [AuditDto.toAbbreviatedJSON(site.getAudits()[0])] }),